Move sunpy.sun.sun to sunpy.coordinates.sun and add deprecated hooks
