+++ sunpy/coordinates/tests/test_sun.py
@@ -5,7 +5,7 @@ from astropy.time import Time
 import astropy.units as u
 from astropy.tests.helper import assert_quantity_allclose
 
-from sunpy.sun import sun
+from sunpy.coordinates import sun
 
 
 @pytest.fixture
